// packages/graph-ui/src/model/types.d.ts
export type NodeId = string;
export type EdgeId = string;
export type GroupId = string;
export interface Point {
    readonly x: number;
    readonly y: number;
}
export interface Size {
    readonly width: number;
    readonly height: number;
}
export interface Rect extends Point, Size {
}
// ...
export interface GraphViewportState {
    readonly x: number;
    readonly y: number;
    readonly zoom: number;
}